requirements.txt
Django>=2.2.4
djangorestframework>=3.10.3
djangorestframework-filters>=1.0.0.dev0